# A couple changes to speedup bulk similarity calculations from Python