--- a/Code/DataStructs/Wrap/wrap_BitOps.cpp
+++ b/Code/DataStructs/Wrap/wrap_BitOps.cpp
@@ -1,6 +1,5 @@
-// $Id$
 //
-//  Copyright (C) 2003-2012 greg Landrum and Rational Discovery LLC
+//  Copyright (C) 2003-2020 greg Landrum and Rational Discovery LLC
 //
 //  @@ All Rights Reserved @@
 //  This file is part of the RDKit.