This makes a minor tweak to speed up calls to `BulkXSimilarity()` by about 10% and introduces new wrapper functions to find the nearest neighbors of a particular fingerprint in a list of other fingerprints.

Here's the performance impact:

| Action | master | this PR |
|----------|-----------|-----------|
| 10K calls to `BulkTanimotoSimilarity()` across ~36K fingerprints | 35.7s | 32.4s |
| find nearest neighbor in the 36K fingerprints for 3K fingerprints using `BulkTanimotoSimilarity()` and a Python sort | 20.0s | 18.7s |
| find nearest neighbor in the 36K fingerprints for 3K fingerprints using `TanimotoSimilarityNeighbors()` | N/A | 7.4s |

There's also a bit of refactoring to simplify `wrap_BitOps.cpp`