@greglandrum 2020-12-10
suggested changes were made, so I'm merging this